server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.1